CM-107: moved core-UI actions into 'redux/ui/thunks'
