+++ src/components/dialogs/UserActiveDialog/UserActiveDialog.jsx
@@ -9,7 +9,7 @@ import React, { PureComponent } from 'react'
 import { Translate } from 'react-redux-i18n'
 import userMonitorService from 'user/monitorService'
 import { connect } from 'react-redux'
-import { logoutAndGoToRoot } from 'redux/core/session/thunks'
+import { logoutAndGoToRoot } from 'redux/ui/thunks'
 import { modalsClose } from 'redux/modals/actions'
 import { stopUserMonitor } from 'redux/ui/actions'
 import ModalDialog from 'components/dialogs/ModalDialog'
