+++ src/components/wallet/WalletWidgetDetail/WalletWidgetDetail.jsx
@@ -17,7 +17,7 @@ import { BLOCKCHAIN_ETHEREUM } from '@chronobank/core/dao/constants'
 import Moment from 'components/common/Moment'
 import WalletModel from '@chronobank/core/models/wallet/WalletModel'
 import MultisigEthWalletModel from '@chronobank/core/models/wallet/MultisigEthWalletModel'
-import { removeWalletAndNavigateToWallets } from 'redux/core/multisigWallet/thunks'
+import { removeWalletAndNavigateToWallets } from 'redux/ui/thunks'
 import SubIconForWallet from '../SubIconForWallet/SubIconForWallet'
 import './WalletWidgetDetail.scss'
 import { prefix } from './lang'
