+++ src/layouts/partials/DrawerMainMenu/MenuTokensList/MenuTokensList.jsx
@@ -13,7 +13,7 @@ import React, { PureComponent } from 'react'
 import { connect } from 'react-redux'
 import { drawerHide, drawerToggle } from 'redux/drawer/actions'
 import { DUCK_SESSION } from '@chronobank/core/redux/session/constants'
-import { logoutAndGoToRoot } from 'redux/core/session/thunks'
+import { logoutAndGoToRoot } from 'redux/ui/thunks'
 import { getBlockchainAddressesList } from '@chronobank/core/redux/session/selectors'
 import { sidesCloseAll, sidesOpen } from 'redux/sides/actions'
 import { MENU_TOKEN_MORE_INFO_PANEL_KEY } from 'redux/sides/constants'
