+++ src/layouts/partials/ProfileContent/ProfileContent.jsx
@@ -10,7 +10,7 @@ import { getNetworkName } from '@chronobank/login/redux/network/thunks'
 import { TOKEN_ICONS } from 'assets'
 import ProfileModel from '@chronobank/core/models/ProfileModel'
 import React, { PureComponent } from 'react'
-import { logoutAndGoToRoot } from 'redux/core/session/thunks'
+import { logoutAndGoToRoot } from 'redux/ui/thunks'
 import { getBlockchainAddressesList } from '@chronobank/core/redux/session/selectors'
 import { modalsOpen } from 'redux/modals/actions'
 import IPFSImage from 'components/common/IPFSImage/IPFSImage'
